requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM integration
anthropic>=0.30.0
openai>=1.0.0

# Discord integration (user account)
discord.py-self>=2.0.0

# IBKR integration
ib-insync>=0.9.86

# Data and analysis
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3

# Logging and monitoring
structlog>=23.1.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
